// app/input-data/types/formTypes.ts
export interface FormData {
  noSuratJalan: string;
  noSeal: string;
  nameDriver: string;
  nameRute: string;
  ruteType: string;
  jenisRitase: string;
  platNomor: string;
  keteranganPlat: string;
  jenisMobil: string;
  jenisTrip: string;
  waktuBerangkat: string;
  waktuTiba: string;
  rateSebelumTax: string;
  ppn: string;
  pph: string;
  totalSetelahTax: string;
  keterangan: string;
  noInvoice: string;
}

export interface InputDataFormProps {
  // Props bisa ditambahkan di sini jika diperlukan
}